Created ModelImportExport script and prefab
