+++ Assets/Prefabs/Game Elements/ModelManager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 94afc50164b53c14c9515ca90bee0de1
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
