+++ Assets/Scenes/Room.unity
@@ -383,7 +383,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 9
+  m_RootOrder: 10
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &435479157
 MonoBehaviour:
@@ -424,6 +424,26 @@ MonoBehaviour:
   sceneViewId: 1
   InstantiationId: 0
   isRuntimeInstantiated: 0
+--- !u!1 &588212834 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 3792769190694460094, guid: 94afc50164b53c14c9515ca90bee0de1,
+    type: 3}
+  m_PrefabInstance: {fileID: 3792769190141423836}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &588212837
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 588212834}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b593c3002bf28274a86c1a0140a4bd6c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  FILE_SERVICE_ENDPOINT: https://fzq7qh0yub.execute-api.us-east-2.amazonaws.com/file
+  NAME_CODE_QUERY_PARAM: nameCode
 --- !u!1 &599916901
 GameObject:
   m_ObjectHideFlags: 0
@@ -916,7 +936,7 @@ Transform:
   - {fileID: 1266323253}
   - {fileID: 900901298}
   m_Father: {fileID: 0}
-  m_RootOrder: 8
+  m_RootOrder: 9
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &2006126369
 GameObject:
@@ -1017,6 +1037,75 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 3b601d5ee9d70b34e86e85fd676fa38c, type: 3}
+--- !u!1001 &3792769190141423836
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 3792769190694460088, guid: 94afc50164b53c14c9515ca90bee0de1,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 8
+      objectReference: {fileID: 0}
+    - target: {fileID: 3792769190694460088, guid: 94afc50164b53c14c9515ca90bee0de1,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3792769190694460088, guid: 94afc50164b53c14c9515ca90bee0de1,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3792769190694460088, guid: 94afc50164b53c14c9515ca90bee0de1,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3792769190694460088, guid: 94afc50164b53c14c9515ca90bee0de1,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3792769190694460088, guid: 94afc50164b53c14c9515ca90bee0de1,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3792769190694460088, guid: 94afc50164b53c14c9515ca90bee0de1,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3792769190694460088, guid: 94afc50164b53c14c9515ca90bee0de1,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3792769190694460088, guid: 94afc50164b53c14c9515ca90bee0de1,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3792769190694460088, guid: 94afc50164b53c14c9515ca90bee0de1,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3792769190694460088, guid: 94afc50164b53c14c9515ca90bee0de1,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3792769190694460094, guid: 94afc50164b53c14c9515ca90bee0de1,
+        type: 3}
+      propertyPath: m_Name
+      value: ModelManager
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 94afc50164b53c14c9515ca90bee0de1, type: 3}
 --- !u!1001 &6885944694426427361
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -1190,21 +1279,3 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 74394877d94ba124799358359acb774c, type: 3}
---- !u!1 &8680544116246475859 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 8680544116335626899, guid: 74394877d94ba124799358359acb774c,
-    type: 3}
-  m_PrefabInstance: {fileID: 8680544116246475858}
-  m_PrefabAsset: {fileID: 0}
---- !u!114 &8680544116246475860
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8680544116246475859}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b593c3002bf28274a86c1a0140a4bd6c, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier:
