+++ Assets/Scripts/ApiRequester.cs
@@ -118,6 +118,7 @@ namespace EasyMeshVR.Web
 
         private IEnumerator RequestPresignedPost()
         {
+            // TODO
             yield return 1;
         }
