+++ Assets/Scripts/ModelImportExport.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9dcbd986d5fcb9e469d1b64085de6219
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
