--- MeshForgeVR-Prototype.sln
@@ -1,20 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Assembly-CSharp", "Assembly-CSharp.csproj", "{B9ACD1F0-02F2-EE85-E560-DA4C0D0509B1}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B9ACD1F0-02F2-EE85-E560-DA4C0D0509B1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B9ACD1F0-02F2-EE85-E560-DA4C0D0509B1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B9ACD1F0-02F2-EE85-E560-DA4C0D0509B1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B9ACD1F0-02F2-EE85-E560-DA4C0D0509B1}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
